requirements.txt
python-dotenv>=1.0.0
pandas>=2.1.0
pyoso>=0.4.0
typer>=0.9.0
pydantic>=2.5.0
requests>=2.31.0